~ address validation: extents, bits, data types, modifiers
Reject addresses whose extent exceeds the 65,536-entry address
space (overflow-safe), bit indices outside the data type's width
or on non-integer types, register-only data types on C/DI areas,
indices without declared dimensions, and malformed or conflicting
byte/word order modifiers.

+++ msd-gateway/src/main/java/com/kevinherron/ignition/modbus/address/ModbusDataType.java
@@ -147,7 +147,7 @@ public sealed interface ModbusDataType
 
     @Override
     public int getRegisterCount() {
-      return (length + 1) / 2;
+      return (int) ((length + 1L) / 2L);
     }
   }
 }
